refactor(schema): move FieldType/FieldSpec from modules/ to dsl/
