+++ crates/limpid/src/check/expr_types.rs
@@ -16,7 +16,7 @@
 use crate::dsl::ast::{BinOp, Expr, ExprKind, UnaryOp, walk_children};
 use crate::dsl::span::Span;
 use crate::functions::{Arity, FunctionRegistry, FunctionSig};
-use crate::modules::schema::{FieldType, type_compatible};
+use crate::dsl::field_schema::{FieldType, type_compatible};
 
 use super::bindings::Bindings;
 use super::suggestions;
