+++ crates/limpid/src/functions/cef/parse.rs
@@ -16,7 +16,7 @@ use anyhow::{Result, bail};
 use crate::functions::primitives::parse_json::{apply_defaults, type_name};
 use crate::functions::primitives::val_to_str;
 use crate::functions::{FunctionRegistry, ParserInfo};
-use crate::modules::schema::{FieldSpec, FieldType};
+use crate::dsl::field_schema::{FieldSpec, FieldType};
 
 pub fn register(reg: &mut FunctionRegistry) {
     reg.register_in("cef", "parse", |arena, args, _event| {
