+++ crates/limpid/src/functions/mod.rs
@@ -54,7 +54,7 @@ use crate::dsl::arena::EventArena;
 use crate::dsl::ast::FunctionDef;
 use crate::dsl::value::Value;
 use crate::event::BorrowedEvent;
-use crate::modules::schema::{FieldSpec, FieldType};
+use crate::dsl::field_schema::{FieldSpec, FieldType};
 
 // ---------------------------------------------------------------------------
 // Function signatures
