+++ crates/limpid/src/functions/primitives/concat.rs
@@ -9,7 +9,7 @@ use anyhow::bail;
 use crate::dsl::arena::EventArena;
 use crate::dsl::value::{ArrayBuilder, Value};
 use crate::functions::{FunctionRegistry, FunctionSig};
-use crate::modules::schema::FieldType;
+use crate::dsl::field_schema::FieldType;
 
 pub fn register(reg: &mut FunctionRegistry) {
     reg.register_with_sig(
