+++ crates/limpid/src/functions/primitives/filter.rs
@@ -5,7 +5,7 @@
 //! loud-fails when the block is omitted).
 use crate::dsl::value::Value;
 use crate::functions::{FunctionRegistry, FunctionSig};
-use crate::modules::schema::FieldType;
+use crate::dsl::field_schema::FieldType;
 
 pub fn register(reg: &mut FunctionRegistry) {
     reg.register_with_sig(
