+++ crates/limpid/src/functions/primitives/find.rs
@@ -8,7 +8,7 @@
 //! generic "expected a block" surprise.
 use crate::dsl::value::Value;
 use crate::functions::{FunctionRegistry, FunctionSig};
-use crate::modules::schema::FieldType;
+use crate::dsl::field_schema::FieldType;
 
 pub fn register(reg: &mut FunctionRegistry) {
     reg.register_with_sig(
