+++ crates/limpid/src/functions/primitives/geoip.rs
@@ -5,7 +5,7 @@
 use super::val_to_str;
 use crate::functions::geoip;
 use crate::functions::{FunctionRegistry, FunctionSig};
-use crate::modules::schema::FieldType;
+use crate::dsl::field_schema::FieldType;
 
 pub fn register(reg: &mut FunctionRegistry) {
     reg.register_with_sig(
