+++ crates/limpid/src/functions/primitives/hashes.rs
@@ -8,7 +8,7 @@ use crate::dsl::value::Value;
 
 use super::val_to_str;
 use crate::functions::{FunctionRegistry, FunctionSig};
-use crate::modules::schema::FieldType;
+use crate::dsl::field_schema::FieldType;
 
 pub fn register(reg: &mut FunctionRegistry) {
     let sig = || FunctionSig::fixed(&[FieldType::String], FieldType::String);
