+++ crates/limpid/src/functions/primitives/hostname.rs
@@ -17,7 +17,7 @@ use std::panic::AssertUnwindSafe;
 
 use crate::dsl::value::Value;
 use crate::functions::{FunctionRegistry, FunctionSig};
-use crate::modules::schema::FieldType;
+use crate::dsl::field_schema::FieldType;
 
 pub fn register(reg: &mut FunctionRegistry) {
     reg.register_with_sig(
