+++ crates/limpid/src/functions/primitives/map.rs
@@ -13,7 +13,7 @@
 //! is `Any` (block bodies are not pinned by FieldType today).
 use crate::dsl::value::Value;
 use crate::functions::{FunctionRegistry, FunctionSig};
-use crate::modules::schema::FieldType;
+use crate::dsl::field_schema::FieldType;
 
 pub fn register(reg: &mut FunctionRegistry) {
     reg.register_with_sig(
