+++ crates/limpid/src/functions/primitives/min.rs
@@ -3,7 +3,7 @@
 //! Sibling of `max` — see `max::pick` for the shared comparison loop.
 
 use crate::functions::{FunctionRegistry, FunctionSig};
-use crate::modules::schema::FieldType;
+use crate::dsl::field_schema::FieldType;
 
 pub fn register(reg: &mut FunctionRegistry) {
     reg.register_with_sig(
