+++ crates/limpid/src/functions/primitives/parse_datetime_rfc2822.rs
@@ -18,7 +18,7 @@ use chrono::Utc;
 use super::val_to_str;
 use crate::dsl::value::Value;
 use crate::functions::{FunctionRegistry, FunctionSig};
-use crate::modules::schema::FieldType;
+use crate::dsl::field_schema::FieldType;
 
 pub fn register(reg: &mut FunctionRegistry) {
     reg.register_with_sig(
