+++ crates/limpid/src/functions/primitives/path.rs
@@ -17,7 +17,7 @@ use anyhow::bail;
 
 use crate::dsl::value::Value;
 use crate::functions::{Arity, FunctionRegistry, FunctionSig};
-use crate::modules::schema::FieldType;
+use crate::dsl::field_schema::FieldType;
 
 pub fn register(reg: &mut FunctionRegistry) {
     // Manual signature: first slot Any (object / null fallthrough),
