+++ crates/limpid/src/functions/primitives/reduce.rs
@@ -5,7 +5,7 @@
 //! stub installs the signature and the missing-block error.
 use crate::dsl::value::Value;
 use crate::functions::{FunctionRegistry, FunctionSig};
-use crate::modules::schema::FieldType;
+use crate::dsl::field_schema::FieldType;
 
 pub fn register(reg: &mut FunctionRegistry) {
     reg.register_with_sig(
