+++ crates/limpid/src/functions/primitives/sum.rs
@@ -10,7 +10,7 @@ use anyhow::bail;
 
 use crate::dsl::value::Value;
 use crate::functions::{FunctionRegistry, FunctionSig};
-use crate::modules::schema::FieldType;
+use crate::dsl::field_schema::FieldType;
 
 pub fn register(reg: &mut FunctionRegistry) {
     reg.register_with_sig(
