+++ crates/limpid/src/functions/primitives/table.rs
@@ -14,7 +14,7 @@ use crate::dsl::value::Value;
 use super::val_to_str;
 use crate::functions::table::TableStore;
 use crate::functions::{FunctionRegistry, FunctionSig};
-use crate::modules::schema::FieldType;
+use crate::dsl::field_schema::FieldType;
 
 pub fn register(reg: &mut FunctionRegistry, table_store: TableStore) {
     {
