+++ crates/limpid/src/functions/primitives/to_json.rs
@@ -18,7 +18,7 @@ use anyhow::{Result, bail};
 
 use crate::dsl::value::Value;
 use crate::functions::{FunctionRegistry, FunctionSig};
-use crate::modules::schema::FieldType;
+use crate::dsl::field_schema::FieldType;
 
 pub fn register(reg: &mut FunctionRegistry) {
     reg.register_with_sig(
