+++ crates/limpid/src/functions/primitives/to_string.rs
@@ -6,7 +6,7 @@ use base64::{Engine as _, engine::general_purpose::STANDARD as B64};
 
 use crate::dsl::value::Value;
 use crate::functions::{FunctionRegistry, FunctionSig};
-use crate::modules::schema::FieldType;
+use crate::dsl::field_schema::FieldType;
 
 pub fn register(reg: &mut FunctionRegistry) {
     reg.register_with_sig(
