+++ crates/limpid/src/functions/syslog/extract_pri.rs
@@ -6,7 +6,7 @@ use crate::dsl::value::Value;
 use crate::functions::primitives::val_to_str;
 use crate::functions::syslog::pri::parse_leading_pri;
 use crate::functions::{FunctionRegistry, FunctionSig};
-use crate::modules::schema::FieldType;
+use crate::dsl::field_schema::FieldType;
 
 pub fn register(reg: &mut FunctionRegistry) {
     reg.register_in_with_sig(
